+++ CHANGES.md
@@ -5,3 +5,4 @@ Changes in v1.4.0
 -----------------
 
 - Added support for "job-retain-until" (Issue #14)
+- Added new PAPPL-Create-Printers operation (Issue #245)